modified:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/controller/RoleController.java 	new file:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/dao/IActionMapDao.java 	new file:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/dao/IMemberMapDao.java 	new file:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/entity/ActionMap.java 	new file:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/entity/MemberMap.java 	modified:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/entity/Role.java 	new file:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/service/IActionMapService.java 	modified:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/service/IActionService.java 	new file:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/service/IMemberMapService.java 	modified:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/service/IRoleService.java 	new file:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/service/impl/ActionMapServiceImpl.java 	new file:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/service/impl/MemberMapServiceImpl.java 	modified:   yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/service/impl/RoleServiceImpl.java 	modified:   yunmiaopu/api/src/main/java/cn/yunmiaopu/user/controller/AccountController.java 	modified:   yunmiaopu/api/src/main/java/cn/yunmiaopu/user/entity/UserSession.java 	modified:   yunmiaopu/api/src/main/java/cn/yunmiaopu/user/util/UserSessionArgumentResolver.java 	modified:   yunmiaopu/api/src/main/java/cn/yunmiaopu/user/util/Users.java 	modified:   yunmiaopu/yunmiaopu.sql

+++ yunmiaopu/api/src/main/java/cn/yunmiaopu/permission/entity/Role.java
@@ -10,18 +10,18 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private  int id;
+    private  long id;
 
     private String name;
 
-    private int creatorUid;
+    private long creatorUid;
 
     private long createTs;
 
     private long updateTs;
 
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -33,11 +33,11 @@ public class Role {
         this.name = name;
     }
 
-    public int getCreatorUid() {
+    public long getCreatorUid() {
         return creatorUid;
     }
 
-    public void setCreatorUid(int creatorUid) {
+    public void setCreatorUid(long creatorUid) {
         this.creatorUid = creatorUid;
     }
